+++ R/subset.R
@@ -263,3 +263,20 @@ st_normalize.stars = function(x, domain = c(0, 0, 1, 1), ...) {
 	}
 	st_stars(x, dimensions = d)
 }
+
+#' @name stars_subset
+#' @param which character or integer; dimension(s) to be flipped
+#' @export
+#' @return \code{st_flip} flips (reverts) the array values along the chosen dimension 
+#' without(s) changing the dimension properties
+st_flip = function(x, which = 1) {
+	if (is.character(which))
+		which = match(which, names(dim(x)))
+	stopifnot(all(which %in% seq_along(dim(x))))
+	dims = lapply(dim(x), seq_len)
+	for (i in which)
+		dims[[ which[i] ]] = rev(dims[[ which[i] ]])
+	for (i in seq_along(x))
+		x[[i]] = do.call(`[`, c(list(x[[i]]), dims))
+	x
+}