+++ tests/testthat/test_ncdf.R
@@ -69,7 +69,8 @@ test_that("euro cordex extra dimvars", {
 
   expect_equal(names(out), c("mask"))
 
-  expect_equal(sf::st_crs(out), sf::st_crs("+proj=lcc +lat_1=30 +lat_2=65 +lat_0=48 +lon_0=9.75 +x_0=-6000 +y_0=-6000 +a=6371229 +b=6371229 +units=m +no_defs"))
+  expect_true(sf::st_crs(out) == 
+  sf::st_crs("+proj=lcc +lat_1=30 +lat_2=65 +lat_0=48 +lon_0=9.75 +x_0=-6000 +y_0=-6000 +a=6371229 +b=6371229 +units=m +no_defs"))
   
   expect_error(out <- suppressWarnings(read_ncdf(f, var = "mask", curvilinear = c("X", "Y"))),
                  "Curvilinear variables not found in file.")